+++ integreat_cms/cms/urls.py
@@ -9,12 +9,14 @@ from django.conf import settings as django_settings
 from django.views.generic import RedirectView
 
 from .forms import LanguageForm, OfferTemplateForm, OrganizationForm, RegionForm
+from .models import Language, OfferTemplate, Organization
 
 from .views import (
     authentication,
     analytics,
     chat,
     dashboard,
+    delete_views,
     events,
     form_views,
     offer_templates,
@@ -178,6 +180,13 @@ urlpatterns = [
                                 ),
                                 name="edit_language",
                             ),
+                            url(
+                                r"^delete$",
+                                delete_views.CustomDeleteView.as_view(
+                                    model=Language,
+                                ),
+                                name="delete_language",
+                            ),
                         ]
                     ),
                 ),
@@ -264,7 +273,9 @@ urlpatterns = [
                             ),
                             url(
                                 r"^delete$",
-                                organizations.organization_actions.delete,
+                                delete_views.CustomDeleteView.as_view(
+                                    model=Organization,
+                                ),
                                 name="delete_organization",
                             ),
                         ]
@@ -324,6 +335,13 @@ urlpatterns = [
                                     form_class=OfferTemplateForm
                                 ),
                                 name="edit_offertemplate",
+                            ),
+                            url(
+                                r"^delete$",
+                                delete_views.CustomDeleteView.as_view(
+                                    model=OfferTemplate, protect_manytomany="regions"
+                                ),
+                                name="delete_offertemplate",
                             ),
                         ]
                     ),