+++ CHANGELOG.md
@@ -19,6 +19,9 @@ UNRELEASED
 * [ [#1134](https://github.com/digitalfabrik/integreat-cms/issues/1134) ] Support legacy XLIFF export for MemoQ WPML filter
 * [ [#943](https://github.com/digitalfabrik/integreat-cms/issues/943) ] Improve performance of content searches
 * [ [#1101](https://github.com/digitalfabrik/integreat-cms/issues/1101) ] Fetch subpages of page tree gradually
+* [ [#1143](https://github.com/digitalfabrik/integreat-cms/issues/1143) ] Hide "Responsible organization" field in page form if no organizations exist
+* [ [#1151](https://github.com/digitalfabrik/integreat-cms/issues/1151) ] Add possibility to delete languages
+* [ [#1106](https://github.com/digitalfabrik/integreat-cms/issues/1106) ] Add possibility to delete offer templates
 
 
 2021.12.0-beta